lab10/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
filelock==3.13.3
platformdirs==4.2.0
sqlparse==0.4.4
virtualenv==20.25.1